limpieza de codigo

+++ src/shared/db/orm.ts
@@ -16,6 +16,7 @@ const clientUrl = process.env.MONGO_URL || 'mongodb://localhost:27017';
 const dbName =
   process.env.TEST_DB_NAME ||
   (process.env.NODE_ENV === 'test' ? 'gym_test' : 'gym');
+const isProduction = process.env.NODE_ENV === 'production';
 
 export const orm = await MikroORM.init({
   driver: MongoDriver,
@@ -32,16 +33,22 @@ export const orm = await MikroORM.init({
   ],
   dbName,
   clientUrl,
-  debug: true,
-  schemaGenerator: {
-    //never in production
-    disableForeignKeys: true,
-    createForeignKeyConstraints: true,
-    ignoreSchema: [],
-  },
+  debug: !isProduction,
+  ...(!isProduction && {
+    schemaGenerator: {
+      // never in production
+      disableForeignKeys: true,
+      createForeignKeyConstraints: true,
+      ignoreSchema: [],
+    },
+  }),
 });
 
 export const syncSchema = async () => {
+  if (isProduction) {
+    return;
+  }
+
   const generator = orm.getSchemaGenerator();
   /*   
   await generator.dropSchema()
